Cleanup code

+++ TbsCore/Helpers/ProxyHelper.cs
@@ -1,10 +1,10 @@
-﻿using HtmlAgilityPack;
-using RestSharp;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Threading.Tasks;
+using HtmlAgilityPack;
+using RestSharp;
 using TbsCore.Models.Access;
 using TbsCore.Models.AccModels;
 using TravBotSharp.Files.Helpers;
@@ -14,12 +14,13 @@ namespace TbsCore.Helpers
     public static class ProxyHelper
     {
         /// <summary>
-        /// Manifest.json string for the chrome extension
+        ///     Manifest.json string for the chrome extension
         /// </summary>
-        private static readonly string manifestJson = "{\"version\": \"1.0.0\",\"manifest_version\": 2,\"name\": \"TBS Proxy Auth Extension\",\"permissions\": [\"proxy\",\"tabs\",\"unlimitedStorage\",\"storage\",\"<all_urls>\",\"webRequest\",\"webRequestBlocking\"],\"background\": {\"scripts\": [\"background.js\"]},\"minimum_chrome_version\":\"22.0.0\"}";
+        private static readonly string manifestJson =
+            "{\"version\": \"1.0.0\",\"manifest_version\": 2,\"name\": \"TBS Proxy Auth Extension\",\"permissions\": [\"proxy\",\"tabs\",\"unlimitedStorage\",\"storage\",\"<all_urls>\",\"webRequest\",\"webRequestBlocking\"],\"background\": {\"scripts\": [\"background.js\"]},\"minimum_chrome_version\":\"22.0.0\"}";
 
         /// <summary>
-        /// Creates chrome extension (.crx) for proxy authentication
+        ///     Creates chrome extension (.crx) for proxy authentication
         /// </summary>
         /// <param name="username">Travian username</param>
         /// <param name="server">Travian server</param>
@@ -35,17 +36,14 @@ namespace TbsCore.Helpers
             CreateFile(Path.Combine(dir, "background.js"), GenerateBackgroundJs(access));
 
             var zipPath = Path.Combine(cacheDir, "chromeExtension.crx");
-            if (File.Exists(zipPath))
-            {
-                File.Delete(zipPath);
-            }
+            if (File.Exists(zipPath)) File.Delete(zipPath);
             ZipFile.CreateFromDirectory(dir, zipPath);
 
             return zipPath;
         }
 
         /// <summary>
-        /// Generates string for the background.js file of the chrome extension
+        ///     Generates string for the background.js file of the chrome extension
         /// </summary>
         /// <param name="access">Extension</param>
         /// <returns>JS code for the extension</returns>
@@ -67,7 +65,8 @@ chrome.proxy.settings.set({value: config, scope: 'regular'}, function() { });
 
 function callbackFn(details)
 {
-    return { authCredentials: { username: '" + access.ProxyUsername.Trim() + @"', password: '" + access.ProxyPassword.Trim() + @"' } };
+    return { authCredentials: { username: '" + access.ProxyUsername.Trim() + @"', password: '" +
+                   access.ProxyPassword.Trim() + @"' } };
 }
 
 chrome.webRequest.onAuthRequired.addListener(
@@ -79,13 +78,13 @@ chrome.webRequest.onAuthRequired.addListener(
         }
 
         /// <summary>
-        /// Create a text file and write to it
+        ///     Create a text file and write to it
         /// </summary>
         /// <param name="path">Path where to create the file</param>
         /// <param name="text">Text to write to the file</param>
         private static void CreateFile(string path, string text)
         {
-            using (StreamWriter writer = File.CreateText(path))
+            using (var writer = File.CreateText(path))
             {
                 writer.Write(text);
             }
@@ -93,20 +92,22 @@ chrome.webRequest.onAuthRequired.addListener(
 
         public static async Task TestProxies(List<Access> access)
         {
-            List<Task> tasks = new List<Task>(access.Count);
+            var tasks = new List<Task>(access.Count);
             access.ForEach(a =>
             {
                 tasks.Add(Task.Run(() =>
                 {
                     var restClient = HttpHelper.InitRestClient(a, "https://api.ipify.org/");
-                    a.Ok = ProxyHelper.TestProxy(restClient, a.Proxy);
+                    a.Ok = TestProxy(restClient, a.Proxy);
                 }));
             });
             await Task.WhenAll(tasks);
         }
 
-        public static bool TestProxy(Account acc) =>
-            TestProxy(acc.Wb.RestClient, acc.Access.GetCurrentAccess().Proxy);
+        public static bool TestProxy(Account acc)
+        {
+            return TestProxy(acc.Wb.RestClient, acc.Access.GetCurrentAccess().Proxy);
+        }
 
         public static bool TestProxy(RestClient client, string proxyIp)
         {
@@ -117,10 +118,10 @@ chrome.webRequest.onAuthRequired.addListener(
             var response = client.Execute(new RestRequest
             {
                 Resource = "",
-                Method = Method.GET,
+                Method = Method.GET
             });
 
-            HtmlDocument doc = new HtmlDocument();
+            var doc = new HtmlDocument();
             doc.LoadHtml(response.Content);
 
             var ip = doc.DocumentNode.InnerText;
